backend/pyproject.toml
[tool.poetry]
name = "messaging-translator"
version = "0.1.0"
description = "A messaging app that translates any sent messages to the target user's target language"
authors = ["Sam Liang <iamsamliang8@gmail.com>"]
license = "Proprietary"
readme = "../README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "~3.11.6"
fastapi = "0.104.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.5.1"}
pydantic-settings = "2.1.0"
python-jose = "3.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
openai = "^1.3.3"
uvicorn = "^0.24.0.post1"
psycopg = {extras = ["c"], version = "^3.1.13"}
alembic = "^1.12.1"
python-multipart = "^0.0.6"
websockets = "^12.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
boto3 = "^1.34.41"
fastapi-mail = "^1.4.1"


[tool.poetry.group.dev.dependencies]
faker = "^20.1.0"
pytest = "^7.4.3"
ipykernel = "^6.26.0"
pytest-dotenv = "^0.5.2"
anyio = "3.7.1"
pytest-asyncio = "^0.21.1"
devtools = "^0.12.2"
mypy = "^1.7.0"
types-python-jose = "^3.3.4.8"
types-passlib = "^1.7.7.13"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
